Script refactor following the comment
