+++ efficientdet_test.py
@@ -4,7 +4,6 @@
 Simple Inference Script of EfficientDet-Pytorch
 """
 import time
-import webcolors
 import torch
 from torch.backends import cudnn
 from matplotlib import colors
@@ -14,9 +13,9 @@ import cv2
 import numpy as np
 
 from efficientdet.utils import BBoxTransform, ClipBoxes
-from utils.utils import preprocess, invert_affine, postprocess
+from utils.utils import preprocess, invert_affine, postprocess, STANDARD_COLORS, standard_to_bgr, get_index_label
 
-compound_coef = 7
+compound_coef = 0
 force_input_size = None  # set None to use default size
 img_path = 'test/img.png'
 
@@ -27,7 +26,7 @@ anchor_scales = [2 ** 0, 2 ** (1.0 / 3.0), 2 ** (2.0 / 3.0)]
 threshold = 0.2
 iou_threshold = 0.2
 
-use_cuda = False
+use_cuda = True
 use_float16 = False
 cudnn.fastest = True
 cudnn.benchmark = True
@@ -42,42 +41,7 @@ obj_list = ['person', 'bicycle', 'car', 'motorcycle', 'airplane', 'bus', 'train'
             'laptop', 'mouse', 'remote', 'keyboard', 'cell phone', 'microwave', 'oven', 'toaster', 'sink',
             'refrigerator', '', 'book', 'clock', 'vase', 'scissors', 'teddy bear', 'hair drier',
             'toothbrush']
-STANDARD_COLORS = [
-    'LawnGreen', 'Chartreuse', 'Aqua','Beige', 'Azure','BlanchedAlmond','Bisque',
-    'Aquamarine', 'BlueViolet', 'BurlyWood', 'CadetBlue', 'AntiqueWhite',
-    'Chocolate', 'Coral', 'CornflowerBlue', 'Cornsilk', 'Crimson', 'Cyan',
-    'DarkCyan', 'DarkGoldenRod', 'DarkGrey', 'DarkKhaki', 'DarkOrange',
-    'DarkOrchid', 'DarkSalmon', 'DarkSeaGreen', 'DarkTurquoise', 'DarkViolet',
-    'DeepPink', 'DeepSkyBlue', 'DodgerBlue', 'FireBrick', 'FloralWhite',
-    'ForestGreen', 'Fuchsia', 'Gainsboro', 'GhostWhite', 'Gold', 'GoldenRod',
-    'Salmon', 'Tan', 'HoneyDew', 'HotPink', 'IndianRed', 'Ivory', 'Khaki',
-    'Lavender', 'LavenderBlush', 'AliceBlue', 'LemonChiffon', 'LightBlue',
-    'LightCoral', 'LightCyan', 'LightGoldenRodYellow', 'LightGray', 'LightGrey',
-    'LightGreen', 'LightPink', 'LightSalmon', 'LightSeaGreen', 'LightSkyBlue',
-    'LightSlateGray', 'LightSlateGrey', 'LightSteelBlue', 'LightYellow', 'Lime',
-    'LimeGreen', 'Linen', 'Magenta', 'MediumAquaMarine', 'MediumOrchid',
-    'MediumPurple', 'MediumSeaGreen', 'MediumSlateBlue', 'MediumSpringGreen',
-    'MediumTurquoise', 'MediumVioletRed', 'MintCream', 'MistyRose', 'Moccasin',
-    'NavajoWhite', 'OldLace', 'Olive', 'OliveDrab', 'Orange', 'OrangeRed',
-    'Orchid', 'PaleGoldenRod', 'PaleGreen', 'PaleTurquoise', 'PaleVioletRed',
-    'PapayaWhip', 'PeachPuff', 'Peru', 'Pink', 'Plum', 'PowderBlue', 'Purple',
-    'Red', 'RosyBrown', 'RoyalBlue', 'SaddleBrown', 'Green', 'SandyBrown',
-    'SeaGreen', 'SeaShell', 'Sienna', 'Silver', 'SkyBlue', 'SlateBlue',
-    'SlateGray', 'SlateGrey', 'Snow', 'SpringGreen', 'SteelBlue', 'GreenYellow',
-    'Teal', 'Thistle', 'Tomato', 'Turquoise', 'Violet', 'Wheat', 'White',
-    'WhiteSmoke', 'Yellow', 'YellowGreen'
-]
 
-def from_colorname_to_bgr(color):
-    rgb_color=webcolors.name_to_rgb(color)
-    result=(rgb_color.blue,rgb_color.green,rgb_color.red)
-    return result
-
-def standard_to_bgr(list_color_name):
-    standard= []
-    for i in range(len(list_color_name)-36): #-36 used to match the len(obj_list)
-        standard.append(from_colorname_to_bgr(list_color_name[i]))
-    return standard
 
 color_list = standard_to_bgr(STANDARD_COLORS)
 # tf bilinear interpolation is different from any other's, just make do
@@ -114,7 +78,6 @@ with torch.no_grad():
                       regressBoxes, clipBoxes,
                       threshold, iou_threshold)
 
-
 def plot_one_box(img, coord, label=None, score=None, color=None, line_thickness=None):
     tl = line_thickness or int(round(0.001 * max(img.shape[0:2])))  # line thickness
     color = color
@@ -128,9 +91,6 @@ def plot_one_box(img, coord, label=None, score=None, color=None, line_thickness=
         cv2.rectangle(img, c1, c2 , color, -1)  # filled
         cv2.putText(img, '{}: {:.0%}'.format(label, score), (c1[0],c1[1] - 2), 0, float(tl) / 3, [0, 0, 0], thickness=tf, lineType=cv2.FONT_HERSHEY_SIMPLEX)
 
-def get_index_label(label):
-    index = int(obj_list.index(label))
-    return index
 
 def display(preds, imgs, imshow=True, imwrite=False):
     for i in range(len(imgs)):
@@ -138,14 +98,15 @@ def display(preds, imgs, imshow=True, imwrite=False):
             continue
 
         for j in range(len(preds[i]['rois'])):
-            x0, y0, x1, y1 = preds[i]['rois'][j].astype(np.int)
+            #x0, y0, x1, y1 = preds[i]['rois'][j].astype(np.int)
+            x1, y1, x2, y2 = preds[i]['rois'][j].astype(np.int)
             obj = obj_list[preds[i]['class_ids'][j]]
             score = float(preds[i]['scores'][j])
-            plot_one_box(imgs[i], [x0, y0, x1, y1], label=obj,score=score,color=color_list[get_index_label(obj)])
+            plot_one_box(imgs[i], [x1, y1, x2, y2], label=obj,score=score,color=color_list[get_index_label(obj, obj_list)])
 
 
         if imshow:
-            cv2.imshow('img',imgs[i])
+            cv2.imshow('img', imgs[i])
             cv2.waitKey(0)
 
         if imwrite:
@@ -153,7 +114,7 @@ def display(preds, imgs, imshow=True, imwrite=False):
 
 
 out = invert_affine(framed_metas, out)
-display(out, ori_imgs, imshow=True, imwrite=False)
+display(out, ori_imgs, imshow=False, imwrite=True)
 
 print('running speed test...')
 with torch.no_grad():
